requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
docstring-parser==0.17.0
eval-type-backport==0.3.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.10.1
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.14.0
livekit==1.1.5
livekit-agents==1.5.2
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.5.2
livekit-plugins-silero==1.5.2
livekit-plugins-turn-detector==1.5.2
livekit-protocol==1.1.5
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.4
onnxruntime==1.24.4
openai==2.31.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pillow==12.2.0
prometheus-client==0.25.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.20.0
pyjwt==2.12.1
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
types-protobuf==7.34.1.20260408
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zipp==3.23.0